+++ CHANGELOG.md
@@ -6,8 +6,11 @@
 #### Added
 - Created player Move state machine
 - Implemented Move SM in the controller
+- Player Controller can return the name of the current states
+- Player GUI
 #### Changed
 - Modified Jump state machine with new states: falling faster and crouching
+- PlayerState is now a pointer inside Player Controller
 
 ### **[0.1.7]** ~ Water _2021-05-26_
 #### Added